refactor(runtime): keep the client-ip policy in the package
`shared/on-cloudflare-edge.ts` exported two things at different altitudes. `onCloudflareEdge` is a
fact about the runtime that both `@lunora/auth` and `@lunora/runtime` need to agree on.
`trustedClientIp` is one package's POLICY on top of it: which header to read, and the decision to
resolve nothing off the edge rather than fall back. A bundler-inlined file the whole repo can
import hands that policy to the next package that reaches for it — and `@lunora/auth` already
disagrees, since it will read an `x-forwarded-for` chain once `trustedProxies` is declared.

`trustedClientIp` moves to `packages/runtime/src/trusted-client-ip.ts`, its only two call sites'
package. It stays internal (not re-exported from the index), so no public API moves.

Its docblock claimed a host behind a proxy it controls "should terminate at something that sets
`cf-connecting-ip` itself". That does not answer the case: a `@lunora/platform-node` origin behind
Cloudflare ALREADY has the header set by the edge, and this code still refuses it, because the code
runs on the origin where `navigator.userAgent` says Node and nothing in the request distinguishes a
fronted deployment from a directly-exposed one. Such a deployment goes from a real per-IP bucket to
one global `no-trusted-ip` bucket any single client can exhaust — a different failure mode, not a
strictly smaller one. Both the docblock and the runtime docs now say so and point those deployments
at an explicit `key`, which is the only place that knows the origin is fronted. No trust-proxy knob
is added; that is a design decision, not a fix.

The remaining `shared/` file's placement paragraph argued against a runtime dependency edge between
`@lunora/auth` and `@lunora/runtime`, which nobody proposed. The real alternative is
`@lunora/platform` — zero-dep leaf, already re-exports a `shared/` file, owns `PlatformCapabilities`
— and the edge to it would be fine. It is rejected on surface instead: platform's index is
published, permanently-supported API, and this is a one-line sniff at an undocumented workerd
detail whose value is being cheap to change; platform is also deliberately declarative, and a
static per-target matrix is a different thing from live host detection. `shared/constant-time-equal`
and `shared/hmac-url` are the precedent. The paragraph now states that, and states the cost of the
choice: the vis build cache does not invalidate on `shared/` edits.
